+++ test/runtests.jl
@@ -1,13 +1,11 @@
 using RelevancePropagation
 
 using Test
-using JuliaFormatter
-using Aqua
 
 @testset verbose = true "RelevancePropagation.jl" begin
     @testset verbose = true "Linting" begin
         @info "Testing linting..."
-        include("test_linting.jl")
+        include("linting.jl")
     end
     @testset "Utilities" begin
         @info "Testing utilities..."
@@ -44,6 +42,10 @@ using Aqua
     @testset "Batches" begin
         @info "Testing analyzers on batches..."
         include("test_batches.jl")
+    end
+    @testset "XAIBase interface" begin
+        @info "Testing XAIBase interface..."
+        include("test_interface.jl")
     end
     @testset "Benchmark correctness" begin
         @info "Testing whether benchmarks are up-to-date..."